requirements.txt
tk
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
matplotlib
numpy
pandas
cartopy
pyserial
